refactor(ipc): extract desktop transport from background service
