+++ apps/browser/src/platform/ipc/transports/index.ts
@@ -1,1 +1,2 @@
 export { WebIpcTransport } from "./web-ipc.transport";
+export { DesktopIpcTransport } from "./desktop-ipc.transport";
